# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Blog",
        "description": "Listado de Endpoints de API Blog",
        "version": "1.0"
    },
    "servers": [
        {
            "url": "http://api.codersfree.test"
        }
    ],
    "paths": {
        "/v1/categories": {
            "get": {
                "tags": [
                    "Categorias"
                ],
                "summary": "Listar todas las Categorias",
                "operationId": "d534950e818920405745e58047885a95",
                "parameters": [
                    {
                        "name": "included",
                        "in": "query",
                        "description": "Incluir una o muchas relaciones entre tablas. Ejemplo: posts o posts.users,posts.tags,posts.image",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "filter[name]",
                        "in": "query",
                        "description": "Filtra los resultados según el nombre de la columna dentro de los corchetes []. El valor 'name' puede ser sutituido por el nombre de cualquier columna de la tabla Posts. Ejemplo: filter[name]=nombreDeCategoria",
                        "required": false,
                        "schema": {
                            "type": "mixed types"
                        }
                    },
                    {
                        "name": "sort",
                        "in": "query",
                        "description": "Ordena los resultados según el valor del parametro. Ejemplo: 'id' para ordernar por el id de forma ascendente o '-id' para ordernar por el id de forma descendente",
                        "required": false,
                        "schema": {
                            "type": "mixed types"
                        }
                    },
                    {
                        "name": "perPage",
                        "in": "query",
                        "description": "Paginar los resultados. Ejemplo: '2' para paginar de 2 en 2",
                        "required": false,
                        "schema": {
                            "type": "number"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "number",
                                                        "example": "1"
                                                    },
                                                    "name": {
                                                        "type": "string",
                                                        "example": "Categoria de prueba"
                                                    },
                                                    "slug": {
                                                        "type": "string",
                                                        "example": "slug-de-prueba"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthenticated."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "access_token": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Categorias"
                ],
                "summary": "Registrar una Categoria",
                "operationId": "c10ce947a02bab958b808bf10246ae28",
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "slug": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "CREATED",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "number",
                                            "example": 1
                                        },
                                        "name": {
                                            "type": "string",
                                            "example": "Categoria de prueba"
                                        },
                                        "slug": {
                                            "type": "string",
                                            "example": "categoria-de-prueba"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Errors",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "The name field is required."
                                        },
                                        "errors": {
                                            "type": "string",
                                            "example": "Objeto de errores"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthenticated."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Forbidden."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "passport": [
                            "create-category"
                        ]
                    }
                ]
            }
        },
...
    },
...
    "components": {
        "securitySchemes": {
            "access_token": {
                "type": "apiKey",
                "description": "Ingrese el token en formato(Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            },
            "passport": {
                "type": "oauth2",
                "description": "Laravel passport oauth2 security.",
                "in": "header",
                "scheme": "https",
                "flows": {
                    "password": {
                        "authorizationUrl": "http://api.codersfree.test/oauth/authorize",
                        "tokenUrl": "http://api.codersfree.test/oauth/token",
                        "refreshUrl": "http://api.codersfree.test/token/refresh",
                        "scopes": {
                            "create-post": "Crear un nuevo post",
                            "read-post": "Leer un post",
                            "update-post": "Actualizar un post",
                            "delete-post": "Eliminar un post",
                            "create-category": "Crear una nuevo categoria",
                            "read-category": "Leer una categoria",
                            "update-category": "Actualizar una categoria",
                            "delete-category": "Eliminar una categoria",
                            "create-tag": "Crear una nuevo etiqueta",
                            "read-tag": "Leer una etiqueta",
                            "update-tag": "Actualizar una etiqueta",
                            "delete-tag": "Eliminar una etiqueta"
                        }
                    }
                }
            }
        }
    },
    "security": [
        {
            "passport": [
                "create-post",
                "read-post",
                "update-post",
                "delete-post",
                "create-category",
                "read-category",
                "update-category",
                "delete-category",
                "create-tag",
                "read-tag",
                "update-tag",
                "delete-tag"
            ]
        }
    ]
}